Separate nfapi_vnf.c/h into LTE and NR implementations
