+++ nfapi/oai_integration/socket/socket_vnf.c
@@ -5,15 +5,15 @@
 
 #include "common/utils/LOG/log.h"
 #include "nfapi.h"
-#include "nfapi_vnf.h"
+#include "vnf/nfapi_nr_vnf.h"
 #include <common/platform_constants.h>
 #include "nfapi/oai_integration/vendor_ext.h" //TODO: Remove this include when removing the Aerial transport stuff
 
 
 void socket_stop_nfapi_p5_p7()
 {
-  get_p7_vnf()->terminate = 1;
-  get_config()->pnf_disconnect_indication = NULL;
+  get_p7_nr_vnf()->terminate = 1;
+  get_nr_config()->pnf_disconnect_indication = NULL;
 }
 
 void socket_nfapi_send_stop_request(vnf_t *vnf)
@@ -292,7 +292,7 @@ static int nfapi_nr_vnf_p5_start(nfapi_vnf_config_t *config)
   struct sctp_initmsg initMsg = {0};
   int noDelay;
 
-  vnf_t *vnf = (vnf_t *)(get_config());
+  vnf_t *vnf = (vnf_t *)(get_nr_config());
 
   NFAPI_TRACE(NFAPI_TRACE_INFO, "Starting P5 VNF connection on port %u\n", config->vnf_p5_port);
 
@@ -746,7 +746,7 @@ static int nfapi_nr_vnf_p7_start(nfapi_vnf_p7_config_t *config)
     return -1;
 
   NFAPI_TRACE(NFAPI_TRACE_INFO, "%s()\n", __FUNCTION__);
-  vnf_p7_t *vnf_p7 = get_p7_vnf();
+  vnf_p7_t *vnf_p7 = get_p7_nr_vnf();
   vnf_p7 = (vnf_p7_t *)config;
 
   // Create p7 receive udp port
